Replace GenServer with TaskSupervisor
* Replace GenServer with TaskSupervisor
* Refactor format to its own module
